% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeOutDirectory.R
\name{changeOutDirectory}
\alias{changeOutDirectory}
\title{Change outDirectory}
\usage{
changeOutDirectory(outdir, create = FALSE)
}
\arguments{
\item{outdir}{directory path}

\item{create}{TRUE/FALSE (will create the directory if it does not exist)}
}
\value{
none (called for side effect: sets outDirectory)
}
\description{
Sets the current outDirectory which will store the Data Ontology and all function output. Option to create directory if does not exist.
}
\examples{
changeOutDirectory(outdir=“~/”, create=FALSE)
}
