% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreConcepts.R
\name{exploreConcepts}
\alias{exploreConcepts}
\title{Extract synonyms and descendants for concepts of interest.}
\usage{
exploreConcepts(vocabulary, codes)
}
\arguments{
\item{vocabulary}{Comma-separated string of relevant vocabularies for inclusion criteria}

\item{codes}{Semi-colon separated string of code concepts for inclusion criteria, corresponding to the order for vocabulary. Multiple codes can be used per vocabulary and should be comma-separated.}
}
\value{
Returns a table of concepts contained under (i.e., below in the heirarchy) the query concept.
}
\description{
For given vocabulary and concept, returns the mapped standard concept(s) as well as decendent concept(s). Requires dataOntology to have been created (makeDataOntology funciton).
}
\examples{
conceptsInfo <- exploreConcepts(vocabulary = “ATC, ICD10CM”, codes = “A01A; K50, K51”)
}
