% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPatients.R
\name{findPatients}
\alias{findPatients}
\title{Find patients based on clinical critera}
\usage{
findPatients(strategy_in = "mapped", vocabulary_in, codes_in,
  function_in = "or", strategy_out = NULL, vocabulary_out = NULL,
  codes_out = NULL, function_out = NULL, declare = FALSE, save = FALSE,
  out_name = NULL)
}
\arguments{
\item{strategy_in}{"mapped" or "direct" (dictates the strategy for how inclusion criteria are treated. "direct" searches for codes as provided, "mapped" maps criteria to standard concepts and finds descendants.}

\item{vocabulary_in}{vocabularies for inclusion criteria (comma-separated string of vocabularies)}

\item{codes_in}{specific concept codes for inclusion criteria (semi-colon separated string of code concepts, corresponding to the order for vocabulary_in. Multiple codes can be used per vocabulary and should be comma-separated.)}

\item{function_in}{"and" or "or" (dictates how multiple inclusion should be treated. "and" necessitates that all inclusion criteria are met (i.e., intersection), while "or" allows for any critera to be met (i.e., union) )}

\item{strategy_out}{"mapped", "direct", or NULL (default) (dictates the strategy for how exclusion are treated. NULL indicates no exclusion criteria.)}

\item{vocabulary_out}{vocabularies for exclusion criteria or NULL (default) (comma-separated string of relevant vocabularies for exclusion criteria. NULL indicates no exclusion criteria)}

\item{codes_out}{specific concept codes for exclusion criteria or NULL (default) (semi-colon separated string of code concepts for inclusion criteria, corresponding to the order for vocabulary_out. Multiple codes can be used per vocabulary and should be comma-separated. NULL indicates no exclusion criteria.)}

\item{function_out}{"and", "or", or NULL (default) (dictates how multiple exclusion should be treated. and necessitates that all exclusion criteria are met (i.e., intersection), while or allows for any critera to be met (i.e., union). NULL indicates no exclusion criteria. )}

\item{declare}{TRUE/FALSE will output status and data information during the process}

\item{save}{TRUE/FALSE whether intermediate components of the search should be saved (e.g., mapped concepts found with unique patient counts per concept).}

\item{out_name}{name assigned to search query or NULL (if save = TRUE, saves query using provided name. If the provided name already exists as a directory (or is NULL), the directory defaults to datetime name)}
}
\value{
List of patients that meet inclusion criteria (and not exclusion criteria if entered).
}
\description{
Identify patients based on clinical data inclusion (and exclusion, if desired) criteria. Flexible to allow for multiple data types, vocabularies, and concepts.
}
\examples{
patient_list = findPatients(strategy_in="mapped", vocabulary_in = "ICD10CM", codes_in = "F41", strategy_out="mapped", vocabulary_out = "MeSH", codes_out = "D002998", function_out = "and")
}
