% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinicalData.R
\name{getClinicalData}
\alias{getClinicalData}
\title{Retrieves all patient clinical data}
\usage{
getClinicalData(patient_list, declare = FALSE)
}
\arguments{
\item{patient_list}{Comma-separated string of patient ids}

\item{declare}{TRUE/FALSE will output status and data information during the process}
}
\value{
a list of tables for each all data within each domain (e.g., Condition) for all patients provided (can access by ptClinicalData$Condition).
}
\description{
Wrapper for domain-specific getData functions (e.g., getObservations). Produces a list of tables for all relevant domains.
}
\examples{
ptClinicalData <- getClinicalData("1,2", declare=TRUE)
}
