% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinicalData.R
\name{getConditions}
\alias{getConditions}
\title{Retrieves all patient clinical data from Condition table}
\usage{
getConditions(patient_list, declare = FALSE)
}
\arguments{
\item{patient_list}{Comma-separated string of patient ids}

\item{declare}{TRUE/FALSE will output status and data information during the process}
}
\value{
a table of relevant clinical data contained with in the 'condition_occurrence' table
}
\description{
Produces a table for relevant concepts contained in the 'condition_occurrence' table mapped through the data ontology for a patient list. Data retrieved include: condition_type, condition_status, etc.
}
\examples{
ptCondData <- getConditions("1,2", declare=TRUE)
}
