% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDemographics.R
\name{getDemographics}
\alias{getDemographics}
\title{Retrieves patient demographic data}
\usage{
getDemographics(patient_list = NULL, declare = FALSE)
}
\arguments{
\item{patient_list}{NULL or comma-separated string of patient ids. A provdied patient_list will restrict search to ids. NULL will return demographic data for all available patients.}

\item{declare}{TRUE/FALSE will output status and data information during the process}
}
\value{
table of mapped demographic concepts for all patients or specific ones for a provided csv formmated string of ids
}
\description{
Compiles demographic data for all patients or a given patient list if provided. Concepts are mapped through the created data ontology. Demographic data are retrieved from 'person' and 'death' tables and include: birthdate, deathdate, gender, ethnicity, and race.
}
\examples{
ptDemo <- getDemographics(patient_list=NULL,declare=TRUE)
}
