% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEncounters.R
\name{getEncounters}
\alias{getEncounters}
\title{Retrieves patient clinical encounter data}
\usage{
getEncounters(patient_list, declare = FALSE)
}
\arguments{
\item{patient_list}{Comma-separated string of patient ids}

\item{declare}{TRUE/FALSE will output status and data information during the process}
}
\value{
table of mapped encounter concepts for specific patients contained in a provided csv formmated string of ids.
}
\description{
Compiles encounter data for a given patient list. Concepts are mapped through the created data ontology. Encounter data are retrieved from visit_occurence table and include: visit_type, encounter_type, etc.
}
\examples{
ptEncs <- getEncounters("1,2,3,4",declare=TRUE)
}
