% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinicalData.R
\name{getMeasurements}
\alias{getMeasurements}
\title{Retrieves all patient clinical data from Measurement table}
\usage{
getMeasurements(patient_list, declare = FALSE)
}
\arguments{
\item{patient_list}{Comma-separated string of patient ids}

\item{declare}{TRUE/FALSE will output status and data information during the process}
}
\value{
a table of relevant clinical data contained with in the 'measurement' table
}
\description{
Produces a table for relevant concepts contained in the 'measurement' table mapped through the data ontology for a patient list. Data retrieved include: measurement_type, value, unit, etc.
}
\examples{
ptMeasData <- getMeasurements("1,2", declare=TRUE)
}
