% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinicalData.R
\name{getMedications}
\alias{getMedications}
\title{Retrieves all patient clinical data from Medications table}
\usage{
getMedications(patient_list, declare = FALSE)
}
\arguments{
\item{patient_list}{Comma-separated string of patient ids}

\item{declare}{TRUE/FALSE will output status and data information during the process}
}
\value{
a table of relevant clinical data contained with in the 'drug_exposure' table
}
\description{
Produces a table for relevant concepts contained in the 'drug_exposure' table mapped through the data ontology for a patient list. Data retrieved include: drug_type, route, etc.
}
\examples{
ptsMedsData <- getMedications("1,2", declare=TRUE)
}
