% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinicalData.R
\name{getObservations}
\alias{getObservations}
\title{Retrieves all patient clinical data from Observations table}
\usage{
getObservations(patient_list, declare = FALSE)
}
\arguments{
\item{patient_list}{Comma-separated string of patient ids}

\item{declare}{TRUE/FALSE will output status and data information during the process}
}
\value{
a table of relevant clinical data contained with in the 'observation' table
}
\description{
Produces a table for relevant concepts contained in the 'observation' table mapped through the data ontology for a patient list. Data retrieved include: observation_type, value, etc.
}
\examples{
ptObsData <- getObservations("1,2", declare=TRUE)
}
