% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataOntology.R
\name{makeDataOntology}
\alias{makeDataOntology}
\title{Creates general data ontology}
\usage{
makeDataOntology(declare = FALSE, store_ontology = FALSE)
}
\arguments{
\item{declare}{TRUE/FALSE will output status and data information during the process}

\item{store_ontology}{TRUE/FALSE If TRUE: will attempt to load .rds file from the current outDirectory; will create and save it if it does not exist. If FALSE: will build table.}
}
\value{
Returns a ontology table dictionary of concepts contained in the 'concept' table.
}
\description{
Creates general data ontology used by all data tables from the concept table. Option to save/load as .rds file.
}
\examples{
\dontrun{
dataOntology <- makeDataOntology(declare=FALSE,store_ontology=TRUE)
}
}
