% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDataTypes.R
\name{showDataTypes}
\alias{showDataTypes}
\title{Shows available data types from the OMOP ontology}
\usage{
showDataTypes()
}
\value{
Returns a table of vocabularies contained within clinical domains: Condition, Observation, Measurement, Device, Procedure, Drug.
}
\description{
Details relevant vocabularies per ontological domain. Requires dataOntology to have been created (makeDataOntology funciton).
}
\examples{
showDataTypes()

}
