% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeDemographics.R
\name{summarizeDemographics}
\alias{summarizeDemographics}
\title{Summarizes patient demographic data}
\usage{
summarizeDemographics(ptDemo)
}
\arguments{
\item{ptDemo}{patient demographics table: ptDemo is the patient demographics object from the getDemographics function output.}
}
\value{
none (called for side effect: prints table)
}
\description{
Summarizes patient demographic data from the getDemographics function.
}
\examples{
\dontrun{
summarizeDemographics(ptDemo)
}
}
